+++ tests/layers_views/views/VivView.spec.js
@@ -64,7 +64,11 @@ export function generateViewTests(ViewType, args, linkedViewIds = []) {
       viewStates[id] = defaultViewState;
     });
     const layers = view.getLayers({
-      props: { loader: [{ type: 'loads' }] },
+      props: {
+        loader: [{ type: 'loads' }],
+        loaderSelection: [{ foo: 'bar' }],
+        transitionFields: ['foo']
+      },
       viewStates
     });
     layers &&