+++ CHANGELOG.md
@@ -4,8 +4,11 @@
 
 ### Added
 
+- `onViewportLoad` prop for `ImageLayer` for handling when the data for the viewport has loaded.
+
 ### Changed
 
+- Fix z-slider broken by transition fields.
 - Upgrade deck.gl to 8.4.0-beta.1 to handle aborting tiles after selection better.
 - Rewrite data loaders as `PixelSource` | `PixelSource[]`. Introduce `ZarrPixelSource` and `TiffPixelSource`
 to support other types of images. Migrate `src/loaders` to TypeScript.